Apply small bug fixes and improvements for Paul Sander's Pong project (#47)
* Remove debugging statement

Oops.

* Fix key remapping bug in InputManager

The input manager wasn't removing an old mapping in some cases, causing the incorrect state for an input to sometimes be read and a remapped key not to function (while the suppo9sedly unmapped key would.)

* Implement a method to cancel capturing a keystroke

* Fix bug where SetKeybindingMenuItem would not properly cancel binding

* Update text handling in SetKeybindingMenuItme's update method

* Simplify key rebinding logic

* Fix bug in calculating the ball's new launch angle after hitting the goal

* Update ball collision logic to fix a certain class of bugs

There are two cases that could cause hangs. The first is where the ball hits the top or bottom of the paddle while the paddle is moving vertically against it with a greater magnitude. This means when the ball's y-velocity is reflected the paddle is still moving faster than it. This is resolved by adding the paddle's y-velocity to the ball to ensure the ball is moving away from it faster than the ball is moving.

The second is when the ball gets pinched between a wall and a paddle. This would cause it to continually try to reflect between the two and become endlessly stuck. Resolved this by just setting the y velocity to zero so it could slide between the other two objects.

* Fix relaunch angle range of the ball

The old value is a bug now that the conversion from degrees to radians has been fixed.

* Add instructions to the title screen

* Fix bug with ball being served in the wrong direction after a player scored